+++ src/Kafka/CreateKafkaInstance/components/StandardKafkaForm.tsx
@@ -2,7 +2,6 @@ import { VFC } from "react";
 import { useTranslation } from "react-i18next";
 
 import {
-  Button,
   Form,
   FormGroup,
   TextInput,
@@ -12,19 +11,20 @@ import {
   Slider,
   Text,
   TextVariants,
-  ButtonVariant,
 } from "@patternfly/react-core";
-import { CloudProvidersTiles, CloudRegionSelect } from "./components";
-import { FormGroupWithPopover } from "../../shared/FormGroupWithPopover";
-import { FormProps } from "./machines/types";
+import { CloudRegionSelect } from "./CloudRegionsSelect";
+import { CloudProvidersTiles } from "./CloudProviderTiles";
+import { FormGroupWithPopover } from "../../../shared/FormGroupWithPopover";
+import { FormProps } from "../machines/types";
 
-export type TrialKafkaFormProps = FormProps & {
+export type StandardKafkaFormProps = FormProps & {
+  streamingUnits: number | undefined;
   /*isTesting flag is temporary for show some contet in storybook, not in productio. 
   It will be remove when actual data will available*/
   isTesting?: boolean;
 };
 
-export const TrialKafkaForm: VFC<TrialKafkaFormProps> = ({
+export const StandardKafkaForm: VFC<StandardKafkaFormProps> = ({
   FORM_ID,
   isNameTaken,
   isNameInvalid,
@@ -43,6 +43,7 @@ export const TrialKafkaForm: VFC<TrialKafkaFormProps> = ({
   disableAZTooltip,
   isDisabledSize,
   size,
+  streamingUnits,
   setSize,
   setRegion,
   setAZ,
@@ -108,16 +109,6 @@ export const TrialKafkaForm: VFC<TrialKafkaFormProps> = ({
           isDisabled={disableControls}
           validated={regionValidation}
         />
-        {isTesting && (
-          <>
-            <Text component={TextVariants.p} className="pf-c-form__helper-text">
-              {t("cloud_region_description")}
-            </Text>
-            <Button variant={ButtonVariant.link} isInline>
-              {t("learn_about_cloud_regions")}
-            </Button>
-          </>
-        )}
       </FormGroup>
       <FormGroup
         label={t("availability_zones")}
@@ -135,7 +126,7 @@ export const TrialKafkaForm: VFC<TrialKafkaFormProps> = ({
             <ToggleGroupItem
               text={t("single")}
               value={"single"}
-              isDisabled={disableControls || !(azOptions?.single === true)}
+              isDisabled={disableControls || azOptions?.single !== true}
               buttonId="single"
               isSelected={az === "single"}
               onChange={() => setAZ("single")}
@@ -152,7 +143,7 @@ export const TrialKafkaForm: VFC<TrialKafkaFormProps> = ({
               text={t("multi")}
               value="multi"
               buttonId="multi"
-              isDisabled={disableControls || !(azOptions?.multi === true)}
+              isDisabled={disableControls || azOptions?.multi !== true}
               isSelected={az === "multi"}
               onChange={() => setAZ("multi")}
             />
@@ -164,7 +155,7 @@ export const TrialKafkaForm: VFC<TrialKafkaFormProps> = ({
           labelHead={t("size")}
           fieldId="streaming-size"
           fieldLabel={t("size")}
-          labelBody={t("trial_kafka_size_description")}
+          labelBody={t("standard_kafka_size_help_content")}
           buttonAriaLabel={t("size_field_aria")}
         >
           <div className="pf-c-input-group pf-u-w-50">
@@ -186,11 +177,8 @@ export const TrialKafkaForm: VFC<TrialKafkaFormProps> = ({
             </span>
           </div>
           <Text component={TextVariants.p} className="pf-c-form__helper-text">
-            {t("trial_kafka_size_description")}
+            {t("standard_kafka_streaming_units", { streamingUnits })}
           </Text>
-          <Button variant={ButtonVariant.link} isInline>
-            {t("learn_about_sizes")}
-          </Button>
         </FormGroupWithPopover>
       )}
     </Form>
