+++ src/Kafka/CreateKafkaInstance/components/index.ts
@@ -3,3 +3,5 @@ export * from "./CloudRegionsSelect";
 export * from "./InstanceInfo";
 export * from "./FormAlerts";
 export * from "./ModalAlerts";
+export * from "./TrialKafkaForm";
+export * from "./StandardKafkaForm";
